refactor(venue): drop map iframe and geocoding API
本番で Vercel から Nominatim へ到達できず /api/places/search が常時 502
だったため、会場の地図機能をまるごと簡素化する。

- Nominatim 連携 API (src/server/routes/places.ts) を削除
- iframe 地図埋め込み (venue-map.tsx) を削除
- events テーブルから latitude / longitude カラムを削除
- venue-field.tsx をサジェスト無しの素のテキスト入力に書き換え
- イベント詳細・招待状・参加導線に「Google マップで開く」リンク
  (venue-link.tsx) を配置

ユーザーは会場へ向かう際に結局 Google Map で経路検索するため、lull 側
で完璧な地図表示を持たず、会場名と住所の確認だけに目的を絞る。
