+++ drizzle/0002_jittery_human_cannonball.sql
@@ -1,0 +1,2 @@
+ALTER TABLE `events` DROP COLUMN `latitude`;--> statement-breakpoint
+ALTER TABLE `events` DROP COLUMN `longitude`;
