+++ src/db/schema.ts
@@ -2,7 +2,6 @@ import { relations } from "drizzle-orm";
 import {
   index,
   integer,
-  real,
   sqliteTable,
   text,
   unique,
@@ -144,8 +143,6 @@ export const events = sqliteTable("events", {
   name: text("name").notNull(),
   venue: text("venue").notNull(),
   address: text("address"),
-  latitude: real("latitude"),
-  longitude: real("longitude"),
   startDatetime: text("start_datetime").notNull(),
   openDatetime: text("open_datetime"),
   status: text("status", { enum: EVENT_STATUSES }).notNull().default("draft"),
