+++ src/lib/queries/members.ts
@@ -52,8 +52,6 @@ export type PerformerInvitationForJoin = {
     name: string;
     venue: string;
     address: string | null;
-    latitude: number | null;
-    longitude: number | null;
     startDatetime: string;
     openDatetime: string | null;
     status: EventStatus;
@@ -137,8 +135,6 @@ export async function getPerformerInvitationByToken(
           name: true,
           venue: true,
           address: true,
-          latitude: true,
-          longitude: true,
           startDatetime: true,
           openDatetime: true,
           status: true,
